fix(aacp): 修改 aacp_mcp 表的端口字段类型并更新 BioCode 配置
- 将 aacp_mcp 表中的 port 字段类型从 TINYINT 修改为 INT，以支持更大的端口值。
- 更新 BioCode.pdma.json 中的 updatedTime 和 baseType、domain 字段，确保数据一致性。

+++ src/main/resources/sql/automatic/V1.7.0115__AACP项目初始化.sql
@@ -12,7 +12,7 @@ CREATE TABLE aacp_mcp(
     `code` VARCHAR(16) NOT NULL  COMMENT '唯一编码' ,
     `network_kind` TINYINT NOT NULL  COMMENT '通信协议 0:HTTP+SSE 1:WS' ,
     `host` VARCHAR(45) NOT NULL  COMMENT '主机' ,
-    `port` TINYINT NOT NULL  COMMENT '端口' ,
+    `port` INT NOT NULL  COMMENT '端口' ,
     `auth_kind` TINYINT NOT NULL  COMMENT '鉴权类型 0:无 1:PSK' ,
     `auth_psk` VARCHAR(2000)   COMMENT '预共享密钥' ,
     `status` TINYINT NOT NULL  COMMENT '状态 0:离线 1:在线' ,
